requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-import-export==3.3.1
django-jazzmin==2.6.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.2.0
numpy==1.26.1
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
